My First Commit

+++ demo-dao-jdbc/src/main/java/model/entities/Department.java
@@ -5,22 +5,22 @@ import java.util.Objects;
 
 public class Department implements Serializable {
 
-    private long id;
+    private int id;
     private String name;
 
     public Department() {
     }
 
-    public Department(long id, String name) {
+    public Department(int id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
